generate parity v2.5.13 fixture
