+++ newsfragments/1753.misc.rst
@@ -1,0 +1,1 @@
+Bump Parity version in tests to v2.5.13
